+++ src/Componets/reactHooks/UseContex.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function UseContex() {
+  return <div>use context</div>;
+}
+
+export default UseContex;
